third_party/blink/web_tests/external/wpt/css/css-typed-om/the-stylepropertymap/properties/text-combine-upright-expected.txt

This is a testharness.js-based test.
[FAIL] 'text-combine-upright' does not support 'digits 3'
  assert_class_string: Unsupported value must be a CSSStyleValue and not one of its subclasses expected "[object CSSStyleValue]" but got "[object Undefined]"
Harness: the test ran to completion.
